Updated comments.

+++ Source/CPP/NuttX/Core/ConsoleUserInterface.cxx
@@ -25,7 +25,7 @@ void ConsoleUserInterface::UpdateDisplayTube(StoreLines &storeLines) const
         char *binary = storeLines[lineNumber].Binary();
         char *disassembled = storeLines[lineNumber].Disassemble();
         printf("%4d: 0x%08x - %32s %-16s ; %d\n", lineNumber, (uint) storeLines[lineNumber].ReverseBits(),
-                                        binary, disassembled, storeLines[lineNumber].GetValue());
+                                        binary, disassembled, (int) storeLines[lineNumber].GetValue());
         delete[] disassembled;
         delete[] binary;
     }
